chore: update to cosmos 4.60.0 and move AllVersionsAndDelete out of preview to stable
