+++ src/WebJobs.Extensions.CosmosDB/Trigger/CosmosDBTriggerAttribute.cs
@@ -131,15 +131,13 @@ namespace Microsoft.Azure.WebJobs
         /// </example>
         public string PreferredLocations { get; set; }
 
-#if PREVIEW
         /// <summary>
         /// Gets or sets the Change Feed mode to use when processing changes. Defaults to <see cref="CosmosDBChangeFeedMode.LatestVersion"/>.
         /// </summary>
         /// <remarks>
-        /// Set to <see cref="CosmosDBChangeFeedMode.AllVersionsAndDeletes"/> to receive all intermediate mutations and delete events (preview feature in Azure Cosmos DB).
+        /// Set to <see cref="CosmosDBChangeFeedMode.AllVersionsAndDeletes"/> to receive all intermediate mutations and delete events.
         /// When using AllVersionsAndDeletes the item type should be wrapped with <see cref="Microsoft.Azure.Cosmos.ChangeFeedItem{T}"/>.
         /// </remarks>
         public CosmosDBChangeFeedMode ChangeFeedMode { get; set; } = CosmosDBChangeFeedMode.LatestVersion;
-#endif
     }
 }
