+++ src/WebJobs.Extensions.CosmosDB/WebJobs.Extensions.CosmosDB.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Azure.Cosmos" Version="3.56.0" />
+    <PackageReference Include="Microsoft.Azure.Cosmos" Version="3.60.0" />
     <PackageReference Include="Microsoft.Azure.WebJobs" Version="3.0.44" />
     <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
     <PackageReference Include="Microsoft.Extensions.Azure" Version="1.13.1" />
@@ -15,7 +15,7 @@
   </ItemGroup>
 
   <ItemGroup Condition="'$(Preview)' == 'true'">
-    <PackageReference Update="Microsoft.Azure.Cosmos" Version="3.57.0-preview.0" />
+    <PackageReference Update="Microsoft.Azure.Cosmos" Version="3.61.0-preview.0" />
   </ItemGroup>
 
 </Project>
