+++ test/WebJobs.Extensions.CosmosDB.Tests/CosmosDBEndToEndTests.cs
@@ -10,6 +10,7 @@ using Microsoft.Azure.Cosmos;
 using Microsoft.Azure.WebJobs.Extensions.Tests;
 using Microsoft.Azure.WebJobs.Extensions.Tests.Common;
 using Microsoft.Azure.WebJobs.Extensions.Tests.Extensions.CosmosDB.Models;
+using Microsoft.Azure.WebJobs.Host.Listeners;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
@@ -23,10 +24,11 @@ namespace Microsoft.Azure.WebJobs.Extensions.CosmosDB.Tests
 {
     // The EndToEnd tests require the AzureWebJobsCosmosDBConnectionString environment variable to be set.
     [Trait("Category", "E2E")]
-    public sealed partial class CosmosDBEndToEndTests(ITestOutputHelper output) : IDisposable
+    public sealed class CosmosDBEndToEndTests(ITestOutputHelper output) : IDisposable
     {
         private const string DatabaseName = "E2EDb";
         private const string CollectionName = "E2ECollection";
+        private const string AllVersionsDeleteCollection = "allVersionsAndDeletes";
         private const string LeaseCollectionName = "leases";
         private readonly TestLoggerProvider _loggerProvider = new();
 
@@ -133,6 +135,46 @@ namespace Microsoft.Azure.WebJobs.Extensions.CosmosDB.Tests
                     await host.StopAsync();
                 }
             }
+        }
+
+        [Fact]
+        public async Task CosmosDBEndToEndChangeFeedModeWrongType()
+        {
+            _loggerProvider.ClearAllLogMessages();
+            using var host = BuildHost(typeof(IncorrectBindingTypeTestClass));
+            using var client = await InitializeDocumentClientAsync(
+                host.Services.GetRequiredService<IConfiguration>(), DatabaseName, AllVersionsDeleteCollection, "/_partitionKey");
+
+            try
+            {
+                await host.StartAsync();
+            }
+            catch (FunctionListenerException ex)
+            {
+                InvalidOperationException inner = Assert.IsType<InvalidOperationException>(ex.InnerException);
+                Assert.StartsWith($"When using ChangeFeedMode.AllVersionsAndDeletes, the trigger binding type must be Microsoft.Azure.Cosmos.ChangeFeedItem<T>.", inner.Message);
+            }
+        }
+
+        [Fact(Skip = "Emulator doesn't appear to work with AllChangesAndDelete")]
+        public async Task CosmosDBEndToEndChangeFeedMode()
+        {
+            _loggerProvider.ClearAllLogMessages();
+            using var host = BuildHost(typeof(AllVersionsAndDeleteTestClass));
+            using var client = await InitializeDocumentClientAsync(
+                host.Services.GetRequiredService<IConfiguration>(), DatabaseName, AllVersionsDeleteCollection, "/_partitionKey");
+
+            await host.StartAsync();
+            await Task.Delay(TimeSpan.FromSeconds(10)); // change feed listener takes a bit to start
+
+            Item item = new() { Id = Guid.NewGuid().ToString(), Text = "Some Text" };
+            var container = client.GetContainer(DatabaseName, AllVersionsDeleteCollection);
+            await container.CreateItemAsync(item, PartitionKey.None);
+
+            IReadOnlyCollection<ChangeFeedItem<Item>> input =
+                await AllVersionsAndDeleteTestClass.Task.WaitAsync(TimeSpan.FromSeconds(30));
+
+            Assert.Single(input, i => i.Current.Id == item.Id);
         }
 
         public void Dispose() => EndToEndTestClass.Reset(null);
@@ -336,6 +378,38 @@ namespace Microsoft.Azure.WebJobs.Extensions.CosmosDB.Tests
                     // Log a message if we see the first item from the earlier delay again
                     log.LogInformation("Saw the first document again!");
                 }
+            }
+        }
+
+        private static class IncorrectBindingTypeTestClass
+        {
+            public static void Trigger(
+                [CosmosDBTrigger(
+                    DatabaseName,
+                    AllVersionsDeleteCollection,
+                    CreateLeaseContainerIfNotExists = true,
+                    LeaseContainerPrefix = "ciIncorrectBindingType",
+                    ChangeFeedMode = CosmosDBChangeFeedMode.AllVersionsAndDeletes)] IReadOnlyCollection<Item> documents) // Not ChangeFeedItem<Item>
+            {
+                // This method is intentionally left blank.
+            }
+        }
+
+        private static class AllVersionsAndDeleteTestClass
+        {
+            private static TaskCompletionSource<IReadOnlyCollection<ChangeFeedItem<Item>>> _triggered = new();
+
+            public static Task<IReadOnlyCollection<ChangeFeedItem<Item>>> Task => _triggered.Task;
+
+            public static void Trigger(
+                [CosmosDBTrigger(
+                    DatabaseName,
+                    AllVersionsDeleteCollection,
+                    CreateLeaseContainerIfNotExists = true,
+                    LeaseContainerPrefix = "ciAllVersionsAndDelete",
+                    ChangeFeedMode = CosmosDBChangeFeedMode.AllVersionsAndDeletes)] IReadOnlyCollection<ChangeFeedItem<Item>> documents)
+            {
+                _triggered.TrySetResult(documents);
             }
         }
     }
