+++ test/WebJobs.Extensions.CosmosDB.Tests/Trigger/CosmosDBTriggerAttributeTests.cs
@@ -34,21 +34,17 @@ namespace Microsoft.Azure.WebJobs.Extensions.CosmosDBTrigger.Tests
             Assert.Equal(defaultLeaseCollectionName, attributeWithNoLeaseSpecified.LeaseContainerName);
             Assert.Equal(databaseName, attributeWithNoLeaseSpecified.LeaseDatabaseName);
 
-#if PREVIEW
             Assert.Equal(CosmosDBChangeFeedMode.LatestVersion, attributeWithNoLeaseSpecified.ChangeFeedMode);
-#endif
 
-            CosmosDBTriggerAttribute attributeWithLeaseSpecified = new(databaseName, collectionName) { LeaseDatabaseName = leaseDatabaseName, LeaseContainerName = leaseCollectionName };
+            CosmosDBTriggerAttribute attributeWithLeaseSpecified= new(databaseName, collectionName) { LeaseDatabaseName = leaseDatabaseName, LeaseContainerName = leaseCollectionName };
 
             Assert.Equal(collectionName, attributeWithLeaseSpecified.ContainerName);
             Assert.Equal(databaseName, attributeWithLeaseSpecified.DatabaseName);
             Assert.Equal(leaseCollectionName, attributeWithLeaseSpecified.LeaseContainerName);
             Assert.Equal(leaseDatabaseName, attributeWithLeaseSpecified.LeaseDatabaseName);
 
-#if PREVIEW
             attributeWithLeaseSpecified.ChangeFeedMode = CosmosDBChangeFeedMode.AllVersionsAndDeletes;
             Assert.Equal(CosmosDBChangeFeedMode.AllVersionsAndDeletes, attributeWithLeaseSpecified.ChangeFeedMode);
-#endif
         }
     }
 }
